Tidy up code in setVersion, refactor most functions to external files
