+++ src/createVersion.js
@@ -1,0 +1,6 @@
+export const createVersion = raw => {
+    return {
+        validRegEx: /^([0-9, .]+)([-+][a-z0-9]+)?$/, // Semantic Versioning 2.0
+        raw
+    }
+}
